Moved PositionEncoding.
